Move editor assets into Editor module and convert to TypeScript
