+++ Derelict/Editor/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "type": "module",
   "scripts": {
-    "build": "tsc -p tsconfig.json",
+    "build": "tsc -p tsconfig.json && cp src/styles.css dist/src/styles.css",
     "test": "npm run build && cd dist && node --test"
   },
   "devDependencies": {
