+++ Derelict/public/index.html
@@ -3,10 +3,10 @@
 <head>
   <meta charset="UTF-8" />
   <title>Derelict Editor</title>
-  <link rel="stylesheet" href="styles.css" />
+  <link rel="stylesheet" href="../Editor/dist/src/styles.css" />
 </head>
 <body>
   <div id="app"></div>
-  <script type="module" src="main.js"></script>
+  <script type="module" src="../Editor/dist/src/main.js"></script>
 </body>
 </html>
